+++ client/components/Templates.js
@@ -1,6 +1,7 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {getTemplatesThunk, createStoryThunk} from '../store'
+import ThumbnailsGrid from './ThumbnailGrid'
 
 class TemplatesContainer extends Component {
   componentDidMount() {
@@ -11,33 +12,6 @@ class TemplatesContainer extends Component {
   }
 }
 
-function ThumbnailsGrid(props) {
-  const {list} = props
-  return (
-    <div className="grid">
-      <h1 className="grid-header">Templates</h1>
-      <div className="grid-thumbnails-container">
-        {list.length &&
-          list.map(item => <Thumbnail item={item} key={item.id} createStory={props.createStory} />)}
-      </div>
-    </div>
-  )
-}
-
-function Thumbnail(props) {
-  const {title, coverImgUrl, description, id} = props.item
-  return (
-    <div className="grid-thumbnail">
-      <div className="grid-thumbnail-title">{title}</div>
-      <img className="grid-thumbnail-image" src={coverImgUrl} />
-      <div className="grid-thumbnail-description">{description}</div>
-      <div className="grid-thumbnail-btn" onClick={() => props.createStory(id)}>
-        Create Story
-      </div>
-      {/* TODO: Attach handler to create story in DB and set it to the state as selected story*/}
-    </div>
-  )
-}
 
 function mapState(state) {
   return {
